upgrade jQuery - with a few more modules excluded in custom build (bug 990351)
